Flatten ModuleGroup
It's now just a module and a flat list of its dependent modules, in
dependency order. We return a (cryptic) error if your dependencies are
cyclic.

Move the Show class to Lam.Primitive because we need it to desugar
string interpolation. Use Lam.Primitive.appendString in string
interpolation to remove the dependency on Lam.Monoid.