Add a simple constraint solver
Add a solver for simple equality constraints based on the paper
Modular Type Inference with Local Assumptions.

Currently we can only solve equality constraints involving type
variables, unification variables and constructors. We need to extend it
to support typeclasses, but this is described in the paper.

We then need to generate these constraints from our AST, extract a
substitution from the solved constraint set, and apply that to the AST
to generate a fully typed program.