Be stricter about tuple patterns
To simplify parsing and printing, we now impose a restriction that tuple
patterns must not be split across newlines. This means that this is a
valid pattern

    (a, b, c)

but this is not

    (a
    , b
    , c)

This makes it a lot easier and faster to parse tuple patterns, and has
the nice upside of making all the property tests pass.