Simplify string interpolation
To simplify the compiler, remove support for automatically converting
values to String in string interpolations. This means any expression
inside #{} must have type String.

Where previously you would write

    "The number is #{4}"

you now write

    "The number is #{showInt 4}"

This is definitely less ergonomic, but allows us to remove the builtin
show function and construct a proper "typeclass" for it. When we add
implicit arguments, we may be able to simplify it to

    "The number is #{4}"

where #{e} roughly desugars to `<> show ? e <>` and we try to infer the
value of ?.

As part of this, we add a new primitive showInt for printing integers.