New typechecker works on basic code
In a deviation from the old typechecker, it doesn't like the use of type
variables in the RHS of a datatype which don't appear on the left, e.g.

  type Functor f = Functor { map : (a -> b) -> f a -> f b }

Where a and b don't appear on the LHS. Haskell also doesn't like this,
and wants you to explicitly quantify them like this:

  type Functor f = Functor { map : forall a b. (a -> b) -> f a -> f b }

This seems sensible to me, so I plan to take that approach.