Slightly improve type errors
When performing subtype checking, bubble the failures up so that the
error is the most general one.

For example, if we are checking the definition

foo : Int
foo = []

We used to get the error

Cannot unify <existential var> with Int

Now we get

Cannot unify forall a. [a] with Int

This is a slight improvement because the first type is something the
user might recognise (the type of []) rather than some internally
generated variable.