Chez: prefix types with $
This prevents name clashes between scheme records and constructors.

e.g. if we define in Kite

    type Show = Show { ... }

then we'll get in Scheme

    (define-record-type Data.Show.Show (fields _tag _0))
    (define Data.Show.Show (lambda (_0) (make-Data.Show.Show 0 _0)))

where the first line is the type definition and the second is the
constructor definition. The constructor name clashes with the name of
the scheme record which is backing the type. So instead we do this:

    (define-record-type $Data.Show.Show (fields _tag _0))
    (define Data.Show.Show (lambda (_0) (make-$Data.Show.Show 0 _0)))