Add compile command
This command takes a path to a Kite file and compiles it to a Chez
program.

Also update the run command to use Chez. The run command now compiles
the given file to a Chez program in a temporary file and then spawns a
scheme process to run that file.