Start to consolidate primitive info
Info on primitive types, functions, foreign calls, operators etc. is
scattered across the codebase. This is a first step in bringing it
together a bit. We replace the Canonical.Primitive module with a Prim
module, and move the primitive type info to Type.Primitive. We also
rename Kite.Primitive to Kite.Prim, because it's a bit shorter.