Allow the typechecker to emit variable solutions
This is preparation for the typechecker to return type-annotated
expressions. When an existential variable is solved, we will emit the
solution via the 'Writer' monad. After typechecker completes, we can
then apply these solutions to type annotations in the resulting
expression to produce concrete types.

This should be more efficient than repeatedly traversing the expressions
to solve each existential separately.

The typechecker doesn't yet emit these solutions - this just lays the
groundwork for it to do so.