WIP: walk an expression with in-scope vars
This function allows you to traverse an expression, editing each node,
in a Reader context that gives you all the in-scope variables for that
node.

We'll use this to implement proof-searches for implicits, which need to
know what variables are in scope (and their types).

It doesn't work yet because we need to refactor patterns to store types
alongside them.