Evalute partial applications correctly
We do this by constructing a PAp with the args we have so far.

We don't yet handle function calls with more than the required number of
arguments, e.g. in

    eq(x, y) = eqInt x y
    returnEq(z) = eq
    main() = returnEq 1 2 3

Here `returnEq` takes one argument but we have given it three. This is
OK because it returns a function of arity 2, so `2` and `3` are actually
arguments to the returned `eq`.